requirement.txt
bs4
flask
keras
matplotlib
nltk
numpy
pandas
pandas-datareader
pillow
scikit-learn
supabase
tensorflow
yahoo_fin